+++ back-end/src/main/java/edu/fpt/ofbs/models/IComboDTO.java
@@ -1,11 +1,11 @@
 package edu.fpt.ofbs.models;
 
 public interface IComboDTO {
-	int getCombo_id();
-	
-	 String getImage_combo_id();
-	
-	 String getCombo_name();
-	
-	 float getCombo_price();
+	long getCombo_id();
+
+	String getImage_combo_id();
+
+	String getCombo_name();
+
+	float getCombo_price();
 }
