+++ back-end/src/main/java/edu/fpt/ofbs/repositories/RestaurantRepository.java
@@ -10,7 +10,7 @@ import edu.fpt.ofbs.entities.Restaurant;
 import edu.fpt.ofbs.models.IRestaurantDTO;
 
 @Repository
-public interface RestaurantRepository extends JpaRepository<Restaurant, Integer>{
+public interface RestaurantRepository extends JpaRepository<Restaurant, Long>{
 	@Query(value = "exec sp_getRestaurantsByType @restaurant_type = ?1", nativeQuery = true)
 	List<IRestaurantDTO> getRestaurantsByType(int type);
 	
@@ -18,5 +18,5 @@ public interface RestaurantRepository extends JpaRepository<Restaurant, Integer>
 	List<IRestaurantDTO> searchRestaurants(int type, String province, String district, String restaurantName);
 	
 	@Query(value = "exec sp_getRestaurantById @restaurant_id = ?1", nativeQuery = true)
-	List<IRestaurantDTO> getRestaurantById(int id);
+	List<IRestaurantDTO> getRestaurantById(long id);
 }
