+++ back-end/src/main/java/edu/fpt/ofbs/repositories/ServiceRepository.java
@@ -10,7 +10,7 @@ import edu.fpt.ofbs.entities.Services;
 import edu.fpt.ofbs.models.IServiceDTO;
 
 @Repository
-public interface ServiceRepository extends JpaRepository<Services, Integer>{
+public interface ServiceRepository extends JpaRepository<Services, Long>{
 	@Query(value = "exec sp_getServicesByRestaurantId @restaurant_id = ?1, @category_id = ?2", nativeQuery = true)
-	List<IServiceDTO> getServicesByRestaurantId(int restaurant_id, int category_id);
+	List<IServiceDTO> getServicesByRestaurantId(long restaurant_id, long category_id);
 }
